session_token and structuring

+++ contact-app/server/src/main/java/com/example/contacts/JdbcRepo.java
@@ -1,9 +1,10 @@
 package com.example.contacts;
 
 import com.example.contacts.model.ContactDetails;
-import com.example.contacts.model.SessionObject;
+import com.example.contacts.model.Sessions;
 import com.example.contacts.model.User;
-import com.example.contacts.model.loginInfo;
+import com.example.contacts.utils.Constants;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.HttpStatus;
@@ -36,113 +37,121 @@ public class JdbcRepo {
     RowMapper<ContactDetails> ContactDetailsRowMapper = (rs, rowNum) -> {
             ContactDetails contact = new ContactDetails();
             contact.setId(rs.getInt("id"));
+            contact.setUserId(rs.getInt("user_id"));
             contact.setName(rs.getString("name"));
             contact.setEmail(rs.getString("email"));
             contact.setScore(rs.getInt("score"));
             return contact;
     };
 
-    RowMapper<SessionObject> SessionObjectRowMapper = (rs, rowNum) -> {
-        SessionObject s = new SessionObject();
-        s.setEmail(rs.getString("user_id"));
-        s.setToken(rs.getString("token"));
-        s.setExpire_time(rs.getTimestamp("expire"));
-        return s;
+    RowMapper<Sessions> SessionObjectRowMapper = (rs, rowNum) -> {
+        Sessions sessionObject = new Sessions();
+        // s.setEmail(rs.getString("email"));
+        sessionObject.setUserId(rs.getInt("user_id"));
+        sessionObject.setToken(rs.getString("session_token"));
+        sessionObject.setExpiryTime(rs.getTimestamp("expiry_time"));
+        return sessionObject;
     };
 
-    public boolean check_auth(String sessionToken)
+    // public String generateSessionToken(String email)
+    // {
+    //     Timestamp current_time = Timestamp.from(Instant.now());
+    //     String token = email + current_time;
+
+    //     Timestamp expire_time = new Timestamp(current_time.getTime());
+
+    //     expire_time.setTime(expire_time.getTime() + ((1 * 60 * 60) * 1000)); // add to constant
+
+    //     SessionObject session_info = new SessionObject(email, token, expire_time);
+    //     BeanPropertySqlParameterSource paramSource = new BeanPropertySqlParameterSource(session_info);
+
+    //     String add_session_token = "insert into sessions (user_id, token, expire) VALUES(:email, :token, :expire_time)";
+
+    //     System.out.println('\n' + token + ' ' + expire_time + '\n');
+
+    //     jdbcTemplate.update(add_session_token, paramSource);
+    //     return token;
+    // }
+
+    public Integer check_auth(String sessionToken)
     {
         System.out.println(sessionToken);
         Timestamp current_time = Timestamp.from(Instant.now());
-        String auth_check = "select * from sessions where token = :token";
-        Map sourceParams = new HashMap<>();
-        sourceParams.put("token", sessionToken);
-        SessionObject obj;
+        String auth_check = "select * from sessions where session_token = :session_token";
+        Map<String, Object> sourceParams = new HashMap<>();
+        sourceParams.put("session_token", sessionToken);
+        Sessions session;
         try {
-             obj = jdbcTemplate.queryForObject(auth_check, sourceParams, SessionObjectRowMapper);
-        }
-        catch (EmptyResultDataAccessException e)
-        {
+            session = jdbcTemplate.queryForObject(auth_check, sourceParams, SessionObjectRowMapper);
+        } catch (EmptyResultDataAccessException e) {
             System.out.println(e);
             throw new ResponseStatusException(HttpStatus.NOT_FOUND);
         }
-        if(obj.getExpire_time().getTime() < current_time.getTime())
-        {
+        if (session.getExpiryTime().getTime() < current_time.getTime()) {
             logout(sessionToken);
-            return false;
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+            // return false;
             //redirect to login
         }
-        return true;
+        return session.getUserId();
     }
-    public String register(loginInfo userRegister)
+
+    public String register(User user)
     {
-        BeanPropertySqlParameterSource paramSource = new BeanPropertySqlParameterSource(userRegister);
-
+        BeanPropertySqlParameterSource paramSource = new BeanPropertySqlParameterSource(user);
         String sql_query = "insert into users(name, email, password) values(:name, :email, :password)";
-
-        String user_exists = "SELECT count(*) FROM users WHERE email = :email LIMIT 1" ;
+        String user_exists = "SELECT count(*) FROM users WHERE email = :email LIMIT 1";
 
         int count = jdbcTemplate.queryForObject(user_exists, paramSource , Integer.class);
-
-        if(count > 0) return "Already Registered";
-
-        jdbcTemplate.update(sql_query,paramSource);
+        if(count > 0) return "Error! User Already Registered";
+        jdbcTemplate.update(sql_query, paramSource);
         return "Registered";
 
         // how to redirect to home page or login page
 
     }
 
-    public String generateSessionToken(String email)
-    {
-        Timestamp current_time = Timestamp.from(Instant.now());
-        String token = email+ current_time;
-
-        Timestamp  expire_time = new Timestamp(current_time.getTime());
-
-        expire_time.setTime(expire_time.getTime() + ((1 * 60 * 60)* 1000)); // add to constant
-
-        SessionObject session_info = new SessionObject(email, token, expire_time);
-        BeanPropertySqlParameterSource paramSource = new BeanPropertySqlParameterSource(session_info);
-
-        String add_session_token = "insert into sessions (user_id, token, expire) VALUES(:email, :token, :expire_time)";
-
-        System.out.println('\n' + token + ' ' + expire_time + '\n');
-
-        jdbcTemplate.update(add_session_token, paramSource);
-        return token;
+    public void addSession(User user, String session_token) {
+        Map<String, Object> params = new HashMap<>();
+        params.put("session_token", session_token);
+        params.put("user_id", user.getId());
+        params.put("expiry_time", new Timestamp(System.currentTimeMillis() + Constants.TIME_30_MIN));
+        System.out.println(params);
+        String sql_query = "INSERT INTO sessions(session_token, user_id, expiry_time) VALUES(:session_token, :user_id, :expiry_time)";
+        jdbcTemplate.update(sql_query, params);
     }
-    public String login(loginInfo info)
+
+    public User login(User user)
     {
-        BeanPropertySqlParameterSource paramSource = new BeanPropertySqlParameterSource(info);
+        BeanPropertySqlParameterSource paramSource = new BeanPropertySqlParameterSource(user);
 
         String sql_query = "select * from users where email = :email and password = :password LIMIT 1";
 
         try {
-            User user = jdbcTemplate.queryForObject(sql_query, paramSource,  UserRowMapper);
+            user = jdbcTemplate.queryForObject(sql_query, paramSource,  UserRowMapper);
         }
         catch (EmptyResultDataAccessException e)
         {
             System.out.println(e);
-            return "Error";
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
         }
-
-        return generateSessionToken(info.getEmail());
-
+        System.out.println(user.getEmail());
+        return user;
     }
 
     public void logout(String sessionToken)
     {
-        String logout_query = "delete from sessions where token = :token";
-        Map sourceParams = new HashMap<>();
-        sourceParams.put("token" , sessionToken);
+        String logout_query = "delete from sessions where session_token = :session_token";
+        Map<String, Object> sourceParams = new HashMap<>();
+        sourceParams.put("session_token" , sessionToken);
         jdbcTemplate.update(logout_query, sourceParams);
     }
 
-    public void insert(ContactDetails contact)
+    public void insert(int userId, ContactDetails contact)
     {
+        contact.setUserId(userId);
         BeanPropertySqlParameterSource paramSource = new BeanPropertySqlParameterSource(contact);
-        String sql_insert = "insert into contacts(id, name, email, score)  VALUES(:id, :name, :email, :score)";
+        String sql_insert = "insert into contacts(user_id, name, email, score)  VALUES(:userId, :name, :email, :score)";
         jdbcTemplate.update(sql_insert, paramSource);
     }
 
@@ -155,19 +164,20 @@ public class JdbcRepo {
     public void update(ContactDetails contact)
     {
         BeanPropertySqlParameterSource paramSource = new BeanPropertySqlParameterSource(contact);
-        String sql_update = "update contacts SET id = :id, name = :name, email = :email, score = :score where id = :id ";
+        String sql_update = "update contacts SET name = :name, email = :email, score = :score where id = :id ";
         jdbcTemplate.update(sql_update, paramSource);
     }
 
-    public void delete(ContactDetails contactID)
+    //check authorisation
+    public void delete(Integer userId, Integer contactId)
     {
+        Map<String, Object> params = new HashMap<>();
+        params.put("user_id", userId);
+        params.put("id", contactId);
 
-        Map params = new HashMap<>();
-        params.put("id", contactID.getId());
+        System.out.println("Printed here" + contactId);
 
-        System.out.println("Printed here" + contactID);
-
-        String sql_delete = "delete from contacts where id = :id ";
+        String sql_delete = "delete from contacts where user_id = :user_id and id = :id";
         jdbcTemplate.update(sql_delete, params);
     }
 }
